requirements.txt
requests~=2.24.0
setuptools~=50.3.2
tqdm~=4.51.0
biopython
pyteomics~=4.3.3